+++ internal/remote/connect.go
@@ -8,9 +8,9 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/zltl/sguala/cli/internal/config"
-	"github.com/zltl/sguala/cli/internal/secret"
-	"github.com/zltl/sguala/cli/internal/sshx"
+	"github.com/zltl/sguala/internal/config"
+	"github.com/zltl/sguala/internal/secret"
+	"github.com/zltl/sguala/internal/sshx"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/term"
 )